backend/auth-v2/requirements.txt
pydantic==2.5.0
psycopg2-binary==2.9.7
PyJWT==2.8.0